CG Sound: bugfixed and enhanced announcer sounds [part 1]. Corrected how (when) some sounds are played. For example the sound when flags are returning are no voice sounds they are a 'klick' sound, but they were treated like the announcer voice. This means they were buffered and played after some delay eventually. This is wromg. Such sounds must be played immediately, because its weird to hear a sound with some delay (with the exeption of real announcer voice sounds). On the other hand the timelimit warnings are played immediately, which is not really necessary because the announcer counts minutes and not seconds, this means it seems to be okay to announce time limit warnings with a delay of max. 1 seconds (even the sound file is about that lenght).
NOTE: Additionally I renamed 'CG_AddBufferedSound' to 'CG_AddBufferedAnnouncerSound' just to make it clear that buffered sounds use CHAN_ANNOUNCER by default.
This renaming is eventually not wanted for baseq3e? I think it depends if we also include our 'priority' buffering.
It is also worth to note that I also renamed the function also for the fraglimit warnings. This means the bahaviour is still the same although these warnings can be considered to be used immediately because in FFA modes with lot of players these warnings can be triggered in areally short amount of time. So possibly merge our 'priority' method.